refactor: Extract fetchPublishedVersions from the upgrade command
Keep the progress spinner at the call site.
